In the old set-up of the NN_model class  the training data was given at initialisation. This was convenient, because then the number of input parameters could be derived from the columns of the x_train_df. This number was necessary to define the input layer of the model. The disadvantage is, that retraining the model with new data becomes more complicated.
Another problem with the old approach is the fixed validation split during model fitting. This makes it difficult to see which part of the dataset the model is actually trained on.

In the new set-up only the column names of the x data are given at initialisation, these are used to determine the architecture of the network, and as reference for the x_data for training, validation and application. The latter means that the columns of the df of the x-data is compared to these column names and an exception is raised when the column names of the x_data do not agree with the column names given at initialisation.

x,y data for training and validation are now an explicit input for the train_model method. This makes it easier to retrain the model,
for example when dealing with a moving training period, and makes it more visible for the user which data is used for the validation.
The scaling factors for x and y data is carried out only during the first training. The same scaling is then applied to all other training sessions and the application.